% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{date_convert}
\alias{date_convert}
\title{Date Conversion File}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3504 rows and 11 columns.
}
\usage{
date_convert
}
\description{
Date Conversion File
}
\keyword{datasets}
