% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_variables}
\alias{find_variables}
\title{Retrieve available columns in processed CRAM output sheets}
\usage{
find_variables(df = NULL)
}
\arguments{
\item{df}{data.frame}
}
\value{
tibble containing the two columns. one column with the names of the columns of input df, and a second column with the column names put into title case
}
\description{
Helper function for obtaining variable names to use in plotting functions, plot_cram and plot_multiple_cram
}
