% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_variables}
\alias{get_variables}
\title{Retrieve available columns in CRAM output sheet of interest}
\usage{
get_variables(
  cram_df,
  out_sheet = "SOUTH_PLATTE",
  start = "1900-01-01",
  end = "2099-01-01"
)
}
\arguments{
\item{cram_df}{data.frame containing Processed CRAM output files, output from process_cram function w/ wide = TRUE}

\item{out_sheet}{character string, name of the output sheet that variables should be returned for, if exact output sheet name is not given a fuzzy match will be returned.}

\item{start}{character string, start date year-month-day}

\item{end}{character string, end date year-month-day}
}
\value{
data.frame containing 5 columns, the last 2 columns being the variables in the output sheet of interest and the type of parameter each variable is.
}
\description{
Helper function for obtaining variable names to use in plotting functions, plot_cram and plot_multiple_cram
}
