% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lookup_table}
\alias{lookup_table}
\title{Variable Lookup Table for CRAM models}
\usage{
lookup_table(model_directory = NULL)
}
\arguments{
\item{model_directory}{dataframe with base_year, model_version, and base_folder columns, indicating the location of a CRAM model folder, a single row from the output dataframe from parse_directory function}
}
\value{
dataframe with columns for output_sheet, model_scenario, variable names, descriptions, and parameters
}
\description{
Variable Lookup Table for CRAM models
}
