% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_directory}
\alias{parse_directory}
\title{Parse Year and Model}
\usage{
parse_directory(base_folder)
}
\arguments{
\item{base_folder}{highest level directory housing all the model folders}
}
\value{
data.frame
}
\description{
Parse Year and Model
}
