% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_cram}
\alias{plot_cram}
\title{Plots desired variables from processed CRAM output sheets}
\usage{
plot_cram(
  cram_df = NULL,
  plot_vars = NULL,
  start = "1900-01-01",
  end = "2099-01-01",
  wrap = FALSE
)
}
\arguments{
\item{cram_df}{data.frame containing Processed CRAM output files, output from process_cram function w/ wide = TRUE}

\item{plot_vars}{character vector containing the names of columns in the processed CRAM dataframe that should be plotted. Default is NULL and will plot the first 2 variables}

\item{start}{character string, start date YYYY-MM-DD}

\item{end}{character string, end date YYYY-MM-DD}

\item{wrap}{logical, whether plot should use facet_wrap or facet_grid for plotting. Default is FALSE, uses facet_grid}
}
\value{
ggplot object
}
\description{
Plots desired variables from processed CRAM output sheets
}
