% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_multi_cram}
\alias{plot_multi_cram}
\title{Title}
\usage{
plot_multi_cram(
  cram_mod1,
  cram_mod2,
  out_sheet = "SOUTH_PLATTE",
  plot_vars = NULL,
  start = "1900-01-01",
  end = "2099-01-01",
  wrap = FALSE
)
}
\arguments{
\item{cram_mod1}{data.frame containing Processed CRAM output files, output from process_cram function w/ wide = TRUE}

\item{cram_mod2}{data.frame containing Processed CRAM output files that should be compared to cram_mod1 input. This will also be the output from process_cram function w/ wide = TRUE}

\item{out_sheet}{character string, name of the output sheet that variables should be returned for, if exact output sheet name is not given a fuzzy match will be returned.}

\item{plot_vars}{character vector containing the names of columns in the processed CRAM output sheet of interest, available columns can be found by using the get_variables function}

\item{start}{character string, start date year-month-day}

\item{end}{character string, end date year-month-day}

\item{wrap}{logical indicating if output plots should use facet_wrap if TRUE and facet_grid if FALSE, default is FALSE}
}
\description{
Title
}
