% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_probs}
\alias{replace_probs}
\title{Remove Problems}
\usage{
replace_probs(df, prob_df)
}
\arguments{
\item{df}{original df}

\item{prob_df}{problem df}
}
\value{
data.frame
}
\description{
Replace problems that arise when reading in tab seperated text file
using readr::read_tsv() then calling problems()
}
