% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_cram}
\alias{plot_cram}
\title{Plots desired variables from processed CRAM output sheets}
\usage{
plot_cram(
  cram_df,
  out_sheet = "SOUTH_PLATTE",
  plot_vars = NULL,
  start = "1900-01-01",
  end = "2099-01-01"
)
}
\arguments{
\item{cram_df}{data.frame containing Processed CRAM output files, output from process_cram function w/ wide = TRUE}

\item{out_sheet}{character string, name of the output sheet that variables should be returned for, if exact output sheet name is not given a fuzzy match will be returned.}

\item{plot_vars}{character vector containing the names of columns in the processed CRAM output sheet of interest, available columns can be found by using the get_variables function}

\item{start}{character string, start date year-month-day}

\item{end}{character string, end date year-month-day}
}
\value{
ggplot object
}
\description{
Plots desired variables from processed CRAM output sheets
}
