% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_cram}
\alias{process_cram}
\title{Process CRAM output}
\usage{
process_cram(model_directory = NULL, return_wide = TRUE)
}
\arguments{
\item{model_directory}{dataframe with base_year, model_version, and base_folder columns, indicating the location of a CRAM model folder, a single row from the output dataframe from parse_directory function}

\item{return_wide}{if TRUE, function returns a wide dataset with a column for each header found in the original OUTPUT_SHEET.txt files, otherwise long dataset w/ multiple rows per timestamp}
}
\value{
a dataframe that has all of the data from the OUTPUT_SHEETS found in the model folder that the function is pointed too.
}
\description{
Process CRAM output
}
