% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_probs}
\alias{replace_probs}
\title{Remove Problems}
\usage{
replace_probs(df, prob_df, verbose = FALSE)
}
\arguments{
\item{df}{original df}

\item{prob_df}{problem df}

\item{verbose}{logical, whether messages should output when function is run. Default is FALSE, no messages are printed}
}
\value{
data.frame
}
\description{
Replace problems that arise when reading in tab seperated text file
using readr::read_tsv() then calling problems()
}
