% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_na_cols}
\alias{rm_na_cols}
\title{Remove dataframe columns with only NA in all rows}
\usage{
rm_na_cols(df)
}
\arguments{
\item{df}{A dataframe expected to have entire columns filled with NA}
}
\value{
data.frame
}
\description{
Remove dataframe columns with only NA in all rows
}
