% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tab_vect}
\alias{tab_vect}
\title{Seperates tab seperated values and format into a dataframe}
\usage{
tab_vect(vect)
}
\arguments{
\item{vect}{a vector of values}
}
\value{
data.frame
}
\description{
Seperates tab seperated values and format into a dataframe
}
